requirements.txt
Flask
firebase-admin
google-generativeai # For Gemini API
